Revert "feature: make struct audio interface public"
This reverts commit b03b43c14d5d726b8567dc032445af3da51cf7d0.
