+++ src/audio.c
@@ -18,7 +18,127 @@
 #include <rem.h>
 #include <baresip.h>
 #include "core.h"
-#include "audio.h"
+
+
+/** Magic number */
+#define MAGIC 0x000a0d10
+#include "magic.h"
+
+
+/**
+ * \page GenericAudioStream Generic Audio Stream
+ *
+ * Implements a generic audio stream. The application can allocate multiple
+ * instances of a audio stream, mapping it to a particular SDP media line.
+ * The audio object has a DSP sound card sink and source, and an audio encoder
+ * and decoder. A particular audio object is mapped to a generic media
+ * stream object. Each audio channel has an optional audio filtering chain.
+ *
+ *<pre>
+ *            write  read
+ *              |    /|\
+ *             \|/    |
+ * .------.   .---------.    .-------.
+ * |filter|<--|  audio  |--->|encoder|
+ * '------'   |         |    |-------|
+ *            | object  |--->|decoder|
+ *            '---------'    '-------'
+ *              |    /|\
+ *              |     |
+ *             \|/    |
+ *         .------. .-----.
+ *         |auplay| |ausrc|
+ *         '------' '-----'
+ *</pre>
+ */
+
+enum {
+	MAX_SRATE       = 48000,  /* Maximum sample rate in [Hz] */
+	MAX_CHANNELS    =     2,  /* Maximum number of channels  */
+	MAX_PTIME       =    60,  /* Maximum packet time in [ms] */
+
+	AUDIO_SAMPSZ    = MAX_SRATE * MAX_CHANNELS * MAX_PTIME / 1000,
+};
+
+
+/**
+ * Audio transmit/encoder
+ *
+ *
+ \verbatim
+
+ Processing encoder pipeline:
+
+ .    .-------.   .-------.   .--------.   .--------.
+ |    |       |   |       |   |        |   |        |
+ |O-->| ausrc |-->| aubuf |-->| aufilt |-->| encode |---> RTP
+ |    |       |   |       |   |        |   |        |
+ '    '-------'   '-------'   '--------'   '--------'
+
+ \endverbatim
+ *
+ */
+struct autx {
+	const struct ausrc *as;       /**< Audio Source module             */
+	struct ausrc_st *ausrc;       /**< Audio Source state              */
+	struct ausrc_prm ausrc_prm;   /**< Audio Source parameters         */
+	const struct aucodec *ac;     /**< Current audio encoder           */
+	struct auenc_state *enc;      /**< Audio encoder state (optional)  */
+	struct aubuf *aubuf;          /**< Packetize outgoing stream       */
+	size_t aubuf_maxsz;           /**< Maximum aubuf size in [bytes]   */
+	volatile bool aubuf_started;  /**< Aubuf was started flag          */
+	struct list filtl;            /**< Audio filters in encoding order */
+	struct mbuf *mb;              /**< Buffer for outgoing RTP packets */
+	char *module;                 /**< Audio source module name        */
+	char *device;                 /**< Audio source device name        */
+	void *sampv;                  /**< Sample buffer                   */
+	uint32_t ptime;               /**< Packet time for sending         */
+	uint64_t ts_ext;              /**< Ext. Timestamp for outgoing RTP */
+	uint32_t ts_base;             /**< First timestamp sent            */
+	uint32_t ts_tel;              /**< Timestamp for Telephony Events  */
+	size_t psize;                 /**< Packet size for sending         */
+	bool marker;                  /**< Marker bit for outgoing RTP     */
+	bool muted;                   /**< Audio source is muted           */
+	int cur_key;                  /**< Currently transmitted event     */
+	enum aufmt src_fmt;           /**< Sample format for audio source  */
+	enum aufmt enc_fmt;           /**< Sample format for encoder       */
+
+	struct {
+		uint64_t aubuf_overrun;
+		uint64_t aubuf_underrun;
+	} stats;
+
+	struct {
+		thrd_t tid;           /**< Audio transmit thread           */
+		RE_ATOMIC bool run;   /**< Audio transmit thread running   */
+	} thr;
+
+	mtx_t *mtx;
+};
+
+
+struct audio_recv;
+
+
+/** Generic Audio stream */
+struct audio {
+	MAGIC_DECL                    /**< Magic number for debugging      */
+	struct autx tx;               /**< Transmit                        */
+	struct audio_recv *aur;       /**< Audio Receiver                  */
+	struct stream *strm;          /**< Generic media stream            */
+	struct telev *telev;          /**< Telephony events                */
+	struct config_audio cfg;      /**< Audio configuration             */
+	bool started;                 /**< Stream is started flag          */
+	bool level_enabled;           /**< Audio level RTP ext. enabled    */
+	bool hold;                    /**< Local hold flag                 */
+	bool conference;              /**< Local conference flag           */
+	uint8_t extmap_aulevel;       /**< ID Range 1-14 inclusive         */
+	audio_event_h *eventh;        /**< Event handler                   */
+	audio_level_h *levelh;        /**< Audio level handler             */
+	audio_err_h *errh;            /**< Audio error handler             */
+	void *arg;                    /**< Handler argument                */
+};
+
 
 /* RFC 6464 */
 static const char *uri_aulevel = "urn:ietf:params:rtp-hdrext:ssrc-audio-level";
